medical_analytics_service/requirements.txt
flask==2.0.1
requests==2.26.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.1
python-dotenv==0.19.0
pytest==6.2.5
pylint==2.9.6 